Rename 'models' to 'user' and make constructor/accessors for User #92
