+++ dams-tests/tests/remote_client.rs
@@ -1,7 +1,7 @@
 use dams::blockchain::Blockchain;
 use dams::keys::{KeyId, KeyMaterial, SharedControl};
-use dams::models::UserId;
 use dams::transaction::TransactionApprovalRequest;
+use dams::user::UserId;
 use dams_remote_client::api::*;
 
 #[test]
